server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
pydantic==2.7.1
python-dotenv==1.0.0
loguru==0.7.2
json-repair==0.25.2
httpx==0.27.0
